goal bugs fixed

See merge request COMP2511/22T2/teams/H18A_JELLY/project!25

+++ src/main/java/dungeonmania/entities/movingentities/Mercenary.java
@@ -19,8 +19,9 @@ public class Mercenary extends Entity implements Moveable, Health {
     private int bribe_amount;
     Position prevPosition;
 
-    public Mercenary(int id, Position position, boolean interactable, boolean collidable, int ally_attack,
-            int ally_defence, int mercenary_attack, int mercenary_health, int bribe_radius, int bribe_amount) {
+    public Mercenary(int id, Position position, boolean interactable, boolean collidable, 
+                    double ally_attack, double ally_defence, double mercenary_attack, 
+                    double mercenary_health, int bribe_radius, int bribe_amount) {
 
         super(id, position, interactable, collidable);
         this.ally_attack = ally_attack;
